Fix #1615

+++ fastai/vision/data.py
@@ -288,7 +288,7 @@ class ImageItemList(ItemList):
         df = pd.read_csv(path/csv_name, header=header)
         return cls.from_df(df, path=path, **kwargs)
 
-    def reconstruct(self, t:Tensor): return Image(t.clamp(min=0,max=1))
+    def reconstruct(self, t:Tensor): return Image(t.float().clamp(min=0,max=1))
 
     def show_xys(self, xs, ys, imgsize:int=4, figsize:Optional[Tuple[int,int]]=None, **kwargs):
         "Show the `xs` (inputs) and `ys` (targets) on a figure of `figsize`."
